+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 
 
 
+## 0.6.1
+1. telegramNotifer 初版
+1. binancecgateway request limit on query_history
+1. bitmexgateway request limit on query_history
+
 ## 0.6.0
 1. dydx 重构
 1. binancec，bitmex POSTONLYLIMIT订单支持